Smart enemy and changes to floats

+++ GameTest/src/main/java/com/adam/GameTest/BasicEnemy.java
@@ -17,7 +17,7 @@ public class BasicEnemy extends GameObject {
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	@Override
@@ -36,7 +36,7 @@ public class BasicEnemy extends GameObject {
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x,(int) y, 16, 16);
 	}
 
 }
